create the app view model god object
