+++ ImageConverter/AppMenu.swift
@@ -2,29 +2,40 @@ import SwiftUI
 
 struct AppMenu: View {
     @Environment(\.openSettings) private var openSettings
-    @EnvironmentObject var settings: SettingsViewModel
-
     
+    // The view now depends on the AppViewModel.
+    @EnvironmentObject var appViewModel: AppViewModel
+
+    // This state is purely for the view's fileImporter presentation. It's fine to keep it here.
     @State private var isImporting = false
-    @State private var isProcessing = false
     
-    @State private var isShowingErrorAlert = false
-    @State private var errorMessage = ""
-
+    // We create a computed binding to determine if the alert should be shown.
+    // This is a clean way to bridge the ViewModel's optional error string to the alert's isPresented boolean.
+    private var isShowingErrorAlert: Binding<Bool> {
+        Binding(
+            get: { appViewModel.errorMessage != nil },
+            set: { _ in
+                // When the alert is dismissed, we clear the error message in the ViewModel.
+                appViewModel.errorMessage = nil
+            }
+        )
+    }
+    
     var body: some View {
         VStack(alignment: .leading, spacing: 10) {
             
             Button("Select Images to Convert...") {
                 self.isImporting = true
             }
-            .disabled(isProcessing)
+            // The disabled state is now driven by the ViewModel's published property.
+            .disabled(appViewModel.isProcessing)
             
             Divider()
             
             Button("Settings...") {
                 openSettings()
             }
-            .disabled(isProcessing)
+            .disabled(appViewModel.isProcessing)
             
             Button("Quit") {
                 NSApplication.shared.terminate(nil)
@@ -36,74 +47,24 @@ struct AppMenu: View {
             allowedContentTypes: [.image],
             allowsMultipleSelection: true
         ) { result in
-            if case .success(let files) = result {
-                process(files: files)
-            } else if case .failure(let error) = result {
-                showError(message: error.localizedDescription)
+            switch result {
+            case .success(let files):
+                // The view simply tells the ViewModel to process the files.
+                // It doesn't know or care how the processing is done.
+                appViewModel.process(files: files)
+            case .failure(let error):
+                // If the file importer itself fails, we can set the error on the ViewModel.
+                appViewModel.errorMessage = error.localizedDescription
             }
         }
-        .alert("An Error Occurred", isPresented: $isShowingErrorAlert) {
-            Button("OK") { }
-        } message: {
-            Text(errorMessage)
-        }
-    }
-    
-    private func showError(message: String) {
-        self.errorMessage = message
-        self.isShowingErrorAlert = true
-    }
-    
-    private func process(files: [URL]) {
-        guard !isProcessing else { return }
-        
-        self.isProcessing = true
-        
-        let panel = NSOpenPanel()
-        panel.title = "Select Output Folder"
-        panel.canChooseFiles = false
-        panel.canChooseDirectories = true
-        panel.canCreateDirectories = true
-        
-        guard panel.runModal() == .OK, let outputDirectory = panel.url else {
-            self.isProcessing = false
-            return
-        }
-        
-        DispatchQueue.global(qos: .userInitiated).async {
-            for file in files {
-                let didStartAccessing = file.startAccessingSecurityScopedResource()
-                defer {
-                    if didStartAccessing {
-                        file.stopAccessingSecurityScopedResource()
-                    }
-                }
-                
-                do {
-                    let format: ImageFormat = (settings.outputFormat == 0) ? .jpeg(quality: 0.85) : .png
-                    let outputData = try ImageConverter.convert(
-                        file: file,
-                        maxDimension: settings.outputSize,
-                        format: format
-                    )
-                    
-                    let originalFilename = file.deletingPathExtension().lastPathComponent
-                    let newFilename = "\(originalFilename)-converted.\(format.fileExtension)"
-                    let outputURL = outputDirectory.appendingPathComponent(newFilename)
-                    try outputData.write(to: outputURL)
-                    
-                } catch {
-                    DispatchQueue.main.async {
-                        showError(message: error.localizedDescription)
-                        self.isProcessing = false
-                    }
-                    return
-                }
+        .alert(
+            "An Error Occurred",
+            isPresented: isShowingErrorAlert, // Use our computed binding
+            actions: { Button("OK") {} },
+            message: {
+                // The message text comes directly from the ViewModel.
+                Text(appViewModel.errorMessage ?? "An unknown error occurred.")
             }
-            
-            DispatchQueue.main.async {
-                self.isProcessing = false
-            }
-        }
+        )
     }
 }
